refactor(visualization): move idToBuilding into features/codeMap
idToBuilding maps node ids to CodeMapBuilding objects (a codeMap rendering
concept) and was a shared service in services/ that imported codeMap while
codeMap imported it back — a services->feature layering smell and one of the
grandfathered codeMap cycles.

- Move services/idToBuilding/idToBuilding.service.ts to
  features/codeMap/idToBuilding.service.ts and point its CodeMapBuilding
  import at the internal ./rendering/codeMapBuilding (no longer via facade).
- Repoint codeMap's internal consumers (arrow, mouseEvent, threeScene) to the
  intra-feature path; expose IdToBuildingService via the codeMap facade for
  the two external consumers (sidebarExplorer, nodeContextMenu).

The codeMap<->idToBuilding dependency is now intra-feature (covered by the
app-wide no-circular warn) instead of a cross-feature edge. Behavior
unchanged. tsc, full Jest suite (378 suites), and dep-cruiser (0 errors) pass.
