+++ visualization/app/codeCharta/features/codeMap/arrow/codeMap.arrow.service.ts
@@ -4,7 +4,7 @@ import { Node, EdgeVisibility } from "../../../codeCharta.model"
 import { ArrowHelper, BufferGeometry, CubicBezierCurve3, Line, LineBasicMaterial, Object3D, Vector3 } from "three"
 import { ColorConverter } from "../../../util/color/colorConverter"
 import { CodeMapBuilding } from "../rendering/codeMapBuilding"
-import { IdToBuildingService } from "../../../services/idToBuilding/idToBuilding.service"
+import { IdToBuildingService } from "../idToBuilding.service"
 import { tap } from "rxjs"
 import { debounce } from "../../../util/debounce"
 import { CodeMapArrowStore } from "../stores/codeMapArrow.store"
