+++ visualization/app/codeCharta/features/codeMap/codeMap.mouseEvent.service.spec.ts
@@ -12,7 +12,7 @@ import { LabelSettingsFacade } from "../../features/labelSettings/facade"
 import { CodeMapTooltipService } from "./codeMap.tooltip.service"
 import { ThreeViewerService } from "./threeViewer/threeViewer.service"
 import { idToNodeSelector } from "../../state/selectors/accumulatedData/idToNode.selector"
-import { IdToBuildingService } from "../../services/idToBuilding/idToBuilding.service"
+import { IdToBuildingService } from "./idToBuilding.service"
 import { setRightClickedNodeData } from "../../state/store/appStatus/rightClickedNodeData/rightClickedNodeData.actions"
 import { State, Store } from "@ngrx/store"
 import { MockStore, provideMockStore } from "@ngrx/store/testing"
