+++ visualization/app/codeCharta/features/codeMap/codeMap.mouseEvent.service.ts
@@ -11,7 +11,7 @@ import { Raycaster, Vector2 } from "three"
 import { LabelSettingsFacade } from "../../features/labelSettings/facade"
 import { CodeMapTooltipService } from "./codeMap.tooltip.service"
 import { ThreeViewerService } from "./threeViewer/threeViewer.service"
-import { IdToBuildingService } from "../../services/idToBuilding/idToBuilding.service"
+import { IdToBuildingService } from "./idToBuilding.service"
 import { tap } from "rxjs"
 import { debounce } from "../../util/debounce"
 import { CodeMapMouseEventStore } from "./stores/codeMapMouseEvent.store"
