+++ visualization/app/codeCharta/features/codeMap/facade.ts
@@ -2,6 +2,7 @@ export { CodeMapComponent } from "./codeMap.component"
 export { CodeMapRenderService } from "./codeMap.render.service"
 export { CodeMapTooltipService } from "./codeMap.tooltip.service"
 export { CodeMapMouseEventService, CursorType } from "./codeMap.mouseEvent.service"
+export { IdToBuildingService } from "./idToBuilding.service"
 export { ThreeSceneService } from "./threeViewer/threeSceneService"
 export { ThreeRendererService } from "./threeViewer/threeRenderer.service"
 export { ThreeMapControlsService } from "./threeViewer/threeMapControls.service"
