+++ visualization/app/codeCharta/features/codeMap/threeViewer/threeSceneService.ts
@@ -5,7 +5,7 @@ import { CodeMapNode, LayoutAlgorithm, Node } from "../../../codeCharta.model"
 import { hierarchy } from "d3-hierarchy"
 import { ColorConverter } from "../../../util/color/colorConverter"
 import { FloorLabelDrawer } from "./floorLabels/floorLabelDrawer"
-import { IdToBuildingService } from "../../../services/idToBuilding/idToBuilding.service"
+import { IdToBuildingService } from "../idToBuilding.service"
 import { ThreeRendererService } from "./threeRenderer.service"
 import { Injectable, OnDestroy } from "@angular/core"
 import { treeMapSize } from "../../../util/algorithm/treeMapLayout/treeMapHelper"
