+++ visualization/app/codeCharta/features/sidebarExplorer/components/explorerTree/explorerTree.component.spec.ts
@@ -4,7 +4,7 @@ import { provideMockStore } from "@ngrx/store/testing"
 import { render } from "@testing-library/angular"
 import { of } from "rxjs"
 import { CodeMapNode, NodeType } from "../../../../codeCharta.model"
-import { IdToBuildingService } from "../../../../services/idToBuilding/idToBuilding.service"
+import { IdToBuildingService } from "../../../../features/codeMap/facade"
 import { defaultState } from "../../../../state/store/state.manager"
 import { CodeMapMouseEventService } from "../../../../features/codeMap/facade"
 import { ThreeRendererService } from "../../../../features/codeMap/facade"
