+++ visualization/app/codeCharta/features/sidebarExplorer/components/explorerTreeLevel/explorerTreeLevel.component.ts
@@ -2,7 +2,7 @@ import { NgClass } from "@angular/common"
 import { ChangeDetectionStrategy, Component, computed, DestroyRef, effect, inject, input, OnInit, signal } from "@angular/core"
 import { toSignal } from "@angular/core/rxjs-interop"
 import { CodeMapNode } from "../../../../codeCharta.model"
-import { IdToBuildingService } from "../../../../services/idToBuilding/idToBuilding.service"
+import { IdToBuildingService } from "../../../../features/codeMap/facade"
 import { CodeMapMouseEventService } from "../../../../features/codeMap/facade"
 import { CodeMapTooltipService } from "../../../../features/codeMap/facade"
 import { ThreeRendererService } from "../../../../features/codeMap/facade"
