+++ visualization/app/codeCharta/features/sidebarExplorer/components/sidebarExplorer/sidebarExplorer.component.spec.ts
@@ -5,7 +5,7 @@ import { render, screen } from "@testing-library/angular"
 import userEvent from "@testing-library/user-event"
 import { of } from "rxjs"
 import { CodeMapNode, NodeType } from "../../../../codeCharta.model"
-import { IdToBuildingService } from "../../../../services/idToBuilding/idToBuilding.service"
+import { IdToBuildingService } from "../../../../features/codeMap/facade"
 import { defaultState } from "../../../../state/store/state.manager"
 import { CodeMapMouseEventService } from "../../../../features/codeMap/facade"
 import { ThreeRendererService } from "../../../../features/codeMap/facade"
